refactor: split movement and ownership transfer for wave 5 slice C (#4125)
Extract movement validation vs civilian apply pipelines and ownership
transfer types/core/orchestration into focused libraries; public barrel
exports unchanged.

Refs #4125
